requirements.txt
numpy~=2.2.6
pillow~=11.3.0
torch~=2.8.0
ultralytics~=8.3.176
pandas~=2.3.1
